pyproject.toml
[tool.poetry]
name = "bothub"
version = "0.1.0"
description = ""
authors = ["João Maia <joaocampo2@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
python-telegram-bot = "^11.1"
simplejson = "^3.16"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
